Move dynamic attributes and each-in iteration out of the always-on paths
`NewTreeBuilder.setDynamicAttribute` was a class method, so every bundle
carried the dynamic attribute classes, property normalization, and URL
sanitization behind it. The logic is now `setDynamicAttribute()` in
`vm/attributes/dynamic.ts`, and the two handlers that set dynamic
attributes import it. A template with only static attributes drops it.

The default iterator checked for `EachInWrapper` and carried the object
and map iterators with it. `toIterator` now tries registered extensions
first, and the `-each-in` helper registers the each-in iterator when it
loads.
