+++ tests/node-vitest/tree-shakability.test.js
@@ -25,6 +25,7 @@ it('[dev] source modules with side effects', () => {
       "@ember/-internals/glimmer/lib/components/input.js",
       "@ember/-internals/glimmer/lib/components/link-to.js",
       "@ember/-internals/glimmer/lib/environment.js",
+      "@ember/-internals/glimmer/lib/helpers/each-in.js",
       "@ember/-internals/metal/lib/property_get.js",
       "@ember/-internals/metal/lib/property_set.js",
       "@ember/-internals/metal/lib/tags.js",
@@ -51,6 +52,7 @@ it('[prod] source modules with side effects', () => {
       "@ember/-internals/glimmer/lib/base-renderer.js",
       "@ember/-internals/glimmer/lib/components/input.js",
       "@ember/-internals/glimmer/lib/environment.js",
+      "@ember/-internals/glimmer/lib/helpers/each-in.js",
       "@ember/-internals/metal/lib/property_get.js",
       "@ember/-internals/metal/lib/property_set.js",
       "@ember/-internals/metal/lib/tags.js",
